requirements.txt
django-ckeditor
django-mapbox-location-field
djangorestframework
django-filter
markdown
django-allauth
django-ckeditor
django-crispy-forms
django-betterforms
django-cors-headers